adds buffers, fixes minor bugs

+++ core/src/graphics/buffers/indexbuffer.cpp
@@ -2,13 +2,13 @@
 
 namespace yengine { namespace graphics {
 
-  IndexBuffer::IndexBuffer(GLfloat * data, GLsizei count)
+  IndexBuffer::IndexBuffer(GLushort* data, GLsizei count)
     : m_Count(count)
   {
     glGenBuffers(1, &m_BufferId);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_BufferId);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLushort), data,
-                 GL_STATIC_DRAW);
+    glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLushort),
+                 data, GL_STATIC_DRAW);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
   }
